# Speed up LeaveOneOutEncoder with vectorization.