Over 400X faster while using less memory.

Store category mappings as DataFrames, then do vectorized apply
with `.map()`.  `fit()` just computes the sum and count of `y` for each
level of each column of `X`.  These mappings are stored as a dict
mapping the column name to a DataFrame with `sum` and `count` columns.
`transform()` then applies the map to each column of `X`, plus a little
vectorized math.

There is a speed/space tradeoff, whether to store the mean of `y`
for each level as well as the sum and count.  This was resolved
in favor of space, recomputing the mean in `transform()`, so that
e.g. pickled Transformers will take less space on disk.

```
rows, cats = 1000000, 1000
X = pd.DataFrame({'x': np.random.randint(0, cats, rows).astype(str)})
y = pd.Series(np.random.rand(rows))

# old
%time LeaveOneOutEncoder().fit_transform(X, y)
CPU times: user 2min 47s, sys: 241 ms, total: 2min 48s
Wall time: 2min 48s

# new
CPU times: user 390 ms, sys: 17.3 ms, total: 408 ms
Wall time: 407 ms
```