@jkleint 2018-10-23
Minor API suggestion for `LeaveOneOutEncoder`: the `randomized` and `sigma` parameters are redundant.  You could get rid of `randomized` and just have `sigma` default to `None`.  If you want randomization, just set `sigma > 0`.