@janmotl 2018-10-23
Thank you. It always makes me happy when the size of the code decreases.

Note however, that the behaviour changed when we encounter a unique value during the training time:
```python
    def test_leave_one_out_unique(self):
        X = pd.DataFrame(data=['1', '2', '2', '2', '3'], columns=['col'])
        y = [1, 0, 1, 0, 1]

        encoder = encoders.LeaveOneOutEncoder(impute_missing=False)
        result = encoder.fit(X, y).transform(X, y)

        self.assertFalse(any(result.isnull()), 'There should not be any missing value')
```
I am not sure we should treat a unique value as missing.

The API change makes sense. Go ahead.