@jkleint 2018-10-23
You're welcome!  Glad to help.

Yeah, I didn't fully understand the algorithm when I was translating it and I think I have a mistake.  I believe the intent is that if there's only one sample for a level, we should use the overall target mean, yes?  And that should happen both when y is provided and when it's not?  That's a little cleaner.  I can do that.