@jkleint 2018-10-23
Fixed the issue with missing values, and added some documentation and comments.