@jkleint 2018-10-23
Also added patch to remove `randomize` arg in favor of `sigma`.