@janmotl 2018-10-24
What if there is a missing value in the input:
```python
    def test_leave_one_out_missing(self):
        X = pd.DataFrame(data=['1', '2', '2', '2', '3', None], columns=['col'])
        y = [1, 0, 1, 0, 1, 0]

        encoder = encoders.LeaveOneOutEncoder(impute_missing=False)
        result_fit = encoder.fit(X, y).transform(X, y)

        self.assertTrue(pd.isna(result_fit['col'][5]), 'we expect NaN or None because impute_missing=False')
```
? Shouldn't we preserve the missing value?