@janmotl 2018-10-24
Note: The suggested change depends on the outcome of #92.